Added image url of product to output. (#3)
* Added image url of product to output.

I needed images of products and spent about 3 hours figuring it out. Image url is one of 4 in the 'images' key. I have chosen the first index wich has the highest resolution.

There is mostlikely a better way of doing this, but this works.

* accounting for no image in response

if there's no image in the response, it will se the `image_url` to `None`